Fix 'head' error and speed up CRLF scan.
This commit contains a fix for the head error that smith was throwing
because the checkmeg() function assumed that $1 must always be a file.
The orignal idea for the check was to make sure that all URLs in the
user's hosts list contained a protocol. meg won't work if you forget
to add http:// or https://.

I also decreased the CRLF scanning time by only testing the special
CRLF payloads using the rawhttp library.